Debug results
Fixed the bug that was making all of the outputs into results.txt be
the same by resetting values for the next iteration of for loop in
DiceRollSimulator.
Changed output of FindAverage.py so that it outputs an integer, because
it is referring to the average number of time needed to roll, which
can’t be a double. Also fixed Usage instructions.
Lastly, updated README.txt to clarify what the result of the program
means.